refactor: Migrate to use /app directory
Just following the upgrade guide: https://beta.nextjs.org/docs/upgrade-guide#migrating-from-pages-to-app

This introduces a new GeneralLayout that serves as a slotted component
for navigation and separates the meta tags from the rest of the content
for each page.
