+++ app/media/head.tsx
@@ -1,0 +1,18 @@
+export default function Head() {
+  return (
+    <>
+      <title>Media and Art Initiatives - Willie Chalmers III</title>
+      <meta
+        name="description"
+        content="Willie makes videos and does other artsty stuff. Check out The Willie Diaries"
+      />
+      <link rel="icon" href="/favicon.ico" />
+
+      <meta property="og:site_name" content="Willie Chalmers III" />
+      <meta property="og:title" content="Media and Art Initiatives" />
+      {/* <meta property="og:image" content="" /> */}
+      <meta property="og:type" content="website" />
+      <meta property="og:url" content="https://williecubed.me/media" />
+    </>
+  );
+}
