--- next.config.js
@@ -1,4 +1,0 @@
-/** @type {import('next').NextConfig} */
-module.exports = {
-  reactStrictMode: true,
-};
