+++ package.json
@@ -11,24 +11,25 @@
     "prepare": "husky install"
   },
   "dependencies": {
-    "next": "13.0.5",
+    "@next/font": "^13.1.1",
+    "next": "13.1.1",
     "react": "18.2.0",
     "react-dom": "18.2.0"
   },
   "devDependencies": {
-    "@types/node": "18.11.9",
-    "@types/react": "18.0.25",
+    "@types/node": "18.11.18",
+    "@types/react": "18.0.26",
     "autoprefixer": "^10.4.13",
     "cross-env": "^7.0.3",
-    "eslint": "8.28.0",
-    "eslint-config-next": "13.0.5",
+    "eslint": "8.30.0",
+    "eslint-config-next": "13.1.1",
     "husky": "^8.0.2",
-    "lint-staged": "^13.0.3",
-    "next-sitemap": "^3.1.32",
-    "postcss": "^8.4.19",
-    "prettier": "^2.8.0",
+    "lint-staged": "^13.1.0",
+    "next-sitemap": "^3.1.43",
+    "postcss": "^8.4.20",
+    "prettier": "^2.8.1",
     "tailwindcss": "^3.2.4",
-    "typescript": "4.9.3"
+    "typescript": "4.9.4"
   },
   "lint-staged": {
     "*.js": "eslint --cache --fix",
